Add simple video and fake router functionality
These videos are from coverr.co and this one in particular has a CC0
license. Plus it was the closest thing I could find to an E-Turbo.

Once this is cleaned up a little better, we can work on animations.

We should probably use the history API to have less ugly URLs.